pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shiny-hunter"
version = "0.1.0"
description = "Automatic shiny hunter for Pokémon Gen 1 (Red/Blue/Yellow) via PyBoy"
requires-python = ">=3.11"
dependencies = [
    "pyboy>=2.7,<3",
    "click>=8.1",
    "rich>=13.7",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
shiny-hunt = "shiny_hunter.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
shiny_hunter = ["macros/*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["needs_rom: requires a real Pokémon ROM in roms/"]
